pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
streamlit = "^1.55.0"
pandas = "^2.2.2"
numpy = "1.26.4"
matplotlib = "^3.9.1"
plotly = "^6.0.0"
scipy = "^1.15.0"
urllib3 = "2.6.3"
gitpython = "3.1.41"
idna = "3.11"
jinja2 = "3.1.5"
pillow = "10.2.0"
protobuf = "6.33.5"
pyarrow = "14.0.1"
tornado = "6.5.4"
certifi = ">=2024.0.0"
bcrypt = "^5.0.0"
psycopg2-binary = "^2.9.11"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"